requirements.txt
torch
transformers
datasets
accelerate
deepspeed
peft
pyyaml
pillow
numpy
pytest
